Disregard line comments in Ledger file
According to ledger-cli documentation, comments start with `;'
(in any column, until EOL). The characters `#|*%' also start
comments until the EOL, but only if they are first on the line.

Block comments (separate lines with the text `comment' to begin,
and `end comment' to finish the block) are currently not supported.